v1.1.3 release

+++ wonton/state/state_manager.h
@@ -60,10 +60,10 @@ StateManager(const MeshWrapper& mesh,
   */
   void add_material_names( const std::unordered_map<std::string, int>& names) {
     for (auto& kv : names) {
-    
+
       // this will update, not overwrite
       material_ids_[kv.first]=kv.second;
-      
+
       // likewise
       material_names_[kv.second] = kv.first;
     }
@@ -127,9 +127,9 @@ StateManager(const MeshWrapper& mesh,
         // add the material to the cell
         cell_materials_[cell].insert(kv.first);
       }
-      
+
       // compute the cell index in material inverse map
-      for (int i = 0; i < kv.second.size(); i++)
+      for (unsigned i = 0; i < kv.second.size(); i++)
         cell_index_in_mat_[kv.first][kv.second[i]]=i;
     }
 
@@ -242,13 +242,13 @@ StateManager(const MeshWrapper& mesh,
     @return A reference to the type_info struct for the field's data type
    */
   const std::type_info& get_data_type(std::string var_name) {
-  
+
   	// get a shared state vector base class pointer and cast it to whatever
   	std::shared_ptr<StateVectorBase> pv = get<StateVectorBase>(var_name);
-		
+
 		// define the data type
 		const std::type_info & info = pv->data_type();
-		
+
 		// get the data type
     return info;
   }
@@ -302,7 +302,7 @@ StateManager(const MeshWrapper& mesh,
                                " already exists in the state manager");
 
     // check that the size is correct
-    if (sv->get_data().size() != mesh_.num_entities(sv->get_kind()))
+    if ((int) sv->get_data().size() != mesh_.num_entities(sv->get_kind()))
       throw std::runtime_error(
             "The added data did not have the same number of elements (" +
             std::to_string(sv->get_data().size()) +") as the number of cells ("+
@@ -580,7 +580,8 @@ StateManager(const MeshWrapper& mesh,
     int n = material_cells_[m].size();
 
     // if the space in the row isn't already allocated, fix this
-    if (material_data.size()  != n) material_data.resize(n);
+    if (material_data.size() != unsigned(n))
+      material_data.resize(n);
 
     *data = material_data.data();
   }
@@ -799,7 +800,6 @@ StateManager(const MeshWrapper& mesh,
   */
   template <class T = double>
   bool shape_is_good(const std::shared_ptr<StateVectorMulti<T>> sv) {
-  
     // get the data shape
     std::unordered_map<int, int> shape{get_material_shape()};
 
@@ -813,13 +813,12 @@ StateManager(const MeshWrapper& mesh,
 	  // this one is tricky, but if we build up the state vector material
 	  // we may start with no data, so for the moment, let's pass this
 	  if (data.size() == 0) return true;
-	  
     // check first that there are the correct number of materials
     if (shape.size() != data.size()) return false;
 
     for (const auto& kv : data) {
       if (shape.find(kv.first) == shape.end() ||
-        shape[kv.first] != kv.second.size()
+        shape[kv.first] != (int) kv.second.size()
       ) return false;
     }
 
@@ -828,14 +827,13 @@ StateManager(const MeshWrapper& mesh,
 
   /*!
     @brief Clear material cells and the inverse map of cell materials
-  */  
+  */
   void clear_material_cells() {
     material_cells_.clear();
     cell_materials_.clear();
   }
 
  protected:
- 
   // underlying mesh reference
   const MeshWrapper& mesh_;
 
@@ -866,7 +864,6 @@ StateManager(const MeshWrapper& mesh,
     material_cells_.clear();
     cell_materials_.clear();
   }
-  
 };
 
 }  // namespace Wonton
